feat(tools-android): add primitives for build and run commands
