+++ .changeset/breezy-mirrors-attend.md
@@ -1,0 +1,7 @@
+---
+"@rnx-kit/cli": patch
+---
+
+Added Android support to the experimental commands for building and running
+apps. Again, this still needs more testing and will not be publicly available
+yet.
