+++ .changeset/thin-eyes-judge.md
@@ -1,0 +1,5 @@
+---
+"@rnx-kit/tools-android": patch
+---
+
+Added primitives for building 'build' and 'run' commands
