+++ packages/cli/src/run.ts
@@ -1,6 +1,7 @@
 import type { Config } from "@react-native-community/cli-types";
 import { rnxBuildCommand } from "./build";
-import type { InputParams } from "./build/apple";
+import type { InputParams } from "./build/types";
+import { runAndroid } from "./run/android";
 import { runIOS } from "./run/ios";
 import { runMacOS } from "./run/macos";
 
@@ -10,9 +11,13 @@ export function rnxRun(
   buildParams: InputParams
 ) {
   switch (buildParams.platform) {
+    case "android":
+      return runAndroid(config, buildParams);
+
     case "ios":
     case "visionos":
       return runIOS(config, buildParams);
+
     case "macos":
       return runMacOS(config, buildParams);
   }
